requirements.txt
feedparser==6.0.11
requests==2.31.0
python-dotenv==1.0.0
openai==1.12.0
groq==0.4.2
beautifulsoup4==4.12.3
apscheduler==3.10.4
